% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotGraph.R
\name{barplotGraph}
\alias{barplotGraph}
\title{create a barplot with different point of view from computed Data}
\usage{
barplotGraph(computedDF = computedDF, parameterOfInterest,
  filteredInstallation = FALSE, groupBy = "Installation", print = T)
}
\arguments{
\item{computedDF}{data of rdf Type from \code{\link{collectData}} function}

\item{parameterOfInterest}{variable to perform the decomposition (can be Installation, Type, Year, Experiments)}

\item{filteredInstallation}{FALSE if you don't want to filter on installations, or the name of the installation(s) (same as \code{\link{installationTable}})}

\item{groupBy}{variable to groupBy and color can be (Installation, Type, Year, Experiments)}

\item{print}{boolean, either to print or save the image}
}
\value{
installation rdfType data
}
\description{
create a barplot with different point of view from computed Data
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
DATA = collectData(INST)
barplotGraph(DATA, parameterOfInterest = "Year", groupBy = "Experiments")
barplotGraph(DATA, parameterOfInterest = "Year", groupBy = "Type")
}
}
