% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectVariable.R
\name{collectVariable}
\alias{collectVariable}
\title{collectVariable from the different installations}
\usage{
collectVariable(inst = NULL, instancesNames, instancesApi)
}
\arguments{
\item{inst}{informations of the installations from \code{\link{installationTable}}}

\item{instancesNames}{the name of the installation(s)}

\item{instancesApi}{the address of the REST API of the installation(s)}
}
\value{
installation rdfType and Experiments Data
}
\description{
collectVariable from the different installations
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
DATA = collectVariable(INST)
}
}
