% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\name{exportData}
\alias{exportData}
\title{Export the data to desired format}
\usage{
exportData(DATA = NULL, format = "csv", rawData = TRUE,
  filename = "file")
}
\arguments{
\item{DATA}{Data of the installations from \code{\link{collectScientificObject}}}

\item{format}{The format desired}

\item{rawData}{Want to download the raw data or the refined ones ?}

\item{filename}{name of the file to download (default is 'file')}
}
\value{
Data in a desired format
}
\description{
Export the data to desired format
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
DATA = collectScientificObject(INST)


}
}
