% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installationTable.R
\name{installationTable}
\alias{installationTable}
\title{Set the list of address and names of installations}
\usage{
installationTable(instancesNames, instancesApi)
}
\arguments{
\item{instancesNames}{the name of the installation(s)}

\item{instancesApi}{the address of the REST API of the installation(s)}
}
\value{
installation
}
\description{
Set the list of address and names of installations
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
}
}
