% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarData.R
\name{radarData}
\alias{radarData}
\title{radarData from the different installations}
\usage{
radarData(DATA = NULL, object = "Installation", variable = "Year")
}
\arguments{
\item{DATA}{Data of the installations from \code{\link{collectScientificObject}}}

\item{object}{The object layer for the radar plot (can be 'Installation', 'Year', 'Experiments', 'Type')}

\item{variable}{The variable on which to explore the objects (can be 'Installation', 'Year', 'Experiments', 'Type')}
}
\value{
Data for scientific objects into a format suitable for radarplot (d3radarR)
}
\description{
radarData from the different installations
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
DATA = collectScientificObject(INST)


}
}
