% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotGraph.R
\name{barplotGraph}
\alias{barplotGraph}
\title{create a barplot with different point of view from computed Data}
\usage{
barplotGraph(computedDF, parameterOfInterest,
  filteredInstallation = FALSE, groupBy = "Installation")
}
\arguments{
\item{computedDF}{data of rdf Type from collectData (\code{\link{collectSensor}}, \code{\link{collectVariable}}, \code{\link{collectScientificObject}}) functions}

\item{parameterOfInterest}{variable to perform the decomposition (can be Installation, Type, Year, Experiments)}

\item{filteredInstallation}{FALSE if you don't want to filter on installations, or the name of the installation(s) (same as \code{\link{installationTable}})}

\item{groupBy}{variable to groupBy and color can be (Installation, Type, Year, Experiments)}
}
\value{
installation rdfType data
}
\description{
create a barplot with different point of view from computed Data
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
DATA = collectScientificObject(INST)
barplotGraph(DATA, parameterOfInterest = "Year", groupBy = "Experiments")
barplotGraph(DATA, parameterOfInterest = "Year", groupBy = "Type")
}
}
