% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectData.R
\name{collectData}
\alias{collectData}
\title{collectData from the different installations}
\usage{
collectData(inst = NULL, instancesNames, instancesApi)
}
\arguments{
\item{inst}{informations of the installations from \code{\link{installationTable}}}

\item{instancesNames}{the name of the installation(s)}

\item{instancesApi}{the address of the REST API of the installation(s)}
}
\value{
installation rdfType and Experiments Data
}
\description{
collectData from the different installations
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
DATA = collectData(INST)
}
}
