% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieGraph.R
\name{pieGraph}
\alias{pieGraph}
\title{create a piechart to vizualize scientific object across the network}
\usage{
pieGraph(collectData, parameterOfInterest, filteredInstallation = FALSE,
  print = T)
}
\arguments{
\item{collectData}{data of rdf Type from \code{\link{collectData}} function}

\item{parameterOfInterest}{variable to perform the decomposition (can be Installation, Type, Year, Experiments)}

\item{filteredInstallation}{name of the installation to focus on}

\item{print}{boolean, either to print or save the image}
}
\value{
piechart of scientific objects colored by the argument.
}
\description{
create a piechart to vizualize scientific object across the network
}
\examples{
\donttest{
INST = installationTable(
           instancesApi = c("opensilex.org/openSilexAPI/rest/"),
           instancesNames = c("opensilexDemo")
       )
DATA = collectData(INST)
pieGraph(DATA, parameterOfInterest = "Type")
}
}
